modularize js files
